Update refresh unit test for manifest changes
A recent change means code for creating a manifest is now
common between basicweb and crud

When a crud app is created from a spec it won't create a manifest unless
bluemix is set to true.

This commit updates the tests to specify "bluemix: true" when running
from a spec.