Fix integration tests with custom generator dir
The model and property integration tests would fail with a mismatched
error message that contained the generator name because Yeoman determines
the first part (before the colon) from the install directory of the
generator.

When you clone the generator into a custom directory and npm link it
as you might for development work, this name will be the custom directory
and might not match the usual name ("swiftserver" in this case).